LLM/백엔드 테스트/requirements.txt
flask==2.3.3
langchain
langchain-openai==0.2.8
sentence-transformers==2.2.2
weaviate-client==3.26.7
python-dotenv==1.0.0
huggingface-hub==0.12.1  # 호환 버전 추가